requirements.txt
canvasapi==3.3.0
python-dotenv==1.0.1
Requests==2.32.3
